requirements.txt
alembic
faker
pandas
blinker
click
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-Login
Flask-WTF
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
psycopg2
pymysql 
bcrypt
pyjwt
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
WTForms
zipp